Classes implementing Iterable are collection like
In the current alpha version of vavr, vavr lists implement the Iterable interface.
I would guess that classes implementing Iterable are somewhat collection like. I'm not sure how big this impact is and maybe there is a reason why the Iterable class was tested for equality instead of isAssignable, but I can't see it. Maybe a comment in the code might help in this case.
However working with custom lists is not possible if they only implement Iterable. So if it is possible to change that, it would be helpful for own list implementations.